changed HP to update continuously while building

+++ Warcraft/app/src/main/java/com/android/ecs160/warcraft/Asset.java
@@ -18,7 +18,7 @@ public class Asset {
     int x; //current x pos
     int y; //current y pos
     EDirection direction;
-    int HP;
+    double HP;
 
     Queue<EAssetAction> commands;
     Queue<CTilePosition> positions;
